rename room stream test with REST API
